refactor: change file name/add new folder for remote ojs client
